No reformatting in <Alt Only="LaTeX" content
Usually PCDATA must be reformatted for LaTeX output, in particular
whitespace must be normalized. With this commit we change this for
<Alt Only="LaTeX">...</Alt> content. This allows to include verbatim
LaTeX code in such an element which may involve LaTeX comments
and empty lines (which mean a paragraph break).

We also state this explicitly in the documentation of the Alt-element,
the previous wording already may have been misunderstood this way.

I don't expect many backward compatibility problems, if any.