Avoid InstallValue for non-plain objects
InstallValue is one of a tiny handful of places calling the GAP kernel
function CLONE_OBJ. This function is rather dangerous, e.g. for *types* it is
not really well-defined, see https://github.com/gap-system/gap/issues/1637.
While I am not aware of any ill-effects of the usage here, I think it is best
to avoid it (and perhaps we can at some point even phase out support for
`InstallValue` used on non-plain objects)