make it testable whether errors occur
GAPDoc tries to create manuals even if `MakeGAPDocDoc` notices errors,
such as missing labels used in cross-references.
This is very good if one wants to produce some preliminary version of
the manuals and then to improve them.
On the other hand, it is useful to detect programmatically whether
errors occur, for example in automatic tests.

The current proposal
- adds support for the global option `GAPDocCollectErrors`;
  if the value of this option is a (mutable) list then information
  about errors that were noticed by `MakeGAPDocDoc` get added to this list,
- introduces an internal function `GAPDocFailure` that gets called in
  order to change the list of errors found;
  this happens in the relevant places where currently an `Info` message
  about an error gets printed

Thus one can use the new feature as follows.
```
errs:= [];
MakeGAPDocDoc( ... : GAPDocCollectErrors:= errs );;
if Length( errs ) <> 0 then
  # call `Error( "..." )` or `GapExitCode( false )` or ...
fi;
```

Once this feature becomes available, it can be used in
- GAP's `doc/make_doc.g`
- the `makedoc.g` files of GAP packages that use GAPDoc,
- the `AutoDoc` calls of GAP packages that use AutoDoc.